+++ libs/labelbox/src/labelbox/schema/export_task.py
@@ -1,9 +1,14 @@
+import json
+import os
+import tempfile
+import warnings
 from abc import ABC, abstractmethod
 from dataclasses import dataclass
 from enum import Enum
 from functools import lru_cache
-import json
 from typing import (
+    TYPE_CHECKING,
+    Any,
     Callable,
     Generic,
     Iterator,
@@ -11,17 +16,13 @@ from typing import (
     Tuple,
     TypeVar,
     Union,
-    TYPE_CHECKING,
-    Any,
 )
 
 import requests
-import tempfile
-import os
+from pydantic import BaseModel
 
 from labelbox.schema.task import Task
 from labelbox.utils import _CamelCaseMixin
-from pydantic import BaseModel
 
 if TYPE_CHECKING:
     from labelbox import Client
@@ -120,6 +121,7 @@ class FileRetrieverStrategy(ABC):  # pylint: disable=too-few-public-methods
             )
         response = requests.get(file_info.file, timeout=30)
         response.raise_for_status()
+        response.encoding = "utf-8"
         assert (
             len(response.content)
             == file_info.offsets.end - file_info.offsets.start + 1